+++ CHANGELOG.md
@@ -126,3 +126,5 @@
   * [FIX]: Fix regression where card table would expand vertically while
     dragging a card
   * [FIX]: Include appUri when creating new MiniReels
+  * [FIX]: Fix an issue where a published MiniReel could still be
+    auto-saved